+++ packages/db/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./bootstrap";
 export * from "./client";
+export { resolveDefaultDbPath, resolveBopoInstanceRoot } from "./default-paths";
 export * from "./repositories";
 export * from "./schema";